+++ test/noahtheduke/splint/runner_test.clj
@@ -29,38 +29,38 @@
        :message "Splint encountered an error: matched"}]
     "(very-special-symbol :do-not-match)")
   (expect-match
-    '[{:rule-name naming/single-segment-namespace
-       :form (ns throw-in-middle)
+    [{:rule-name 'naming/single-segment-namespace
+       :form '(ns throw-in-middle)
        :message "throw-in-middle is a single segment. Consider adding an additional segment."
        :alt nil
        :line 5
        :column 1
        :end-row 5
        :end-col 21
-       :filename "corpus/throw_in_middle.clj"}
-      {:rule-name dev/throws-on-match
-       :form (very-special-symbol :do-not-match)
+       :filename (io/file "corpus/throw_in_middle.clj")}
+      {:rule-name 'dev/throws-on-match
+       :form '(very-special-symbol :do-not-match)
        :message "Splint encountered an error: matched"
        :alt nil
        :line 7
        :column 1
        :end-row 7
        :end-col 36
-       :filename "corpus/throw_in_middle.clj"}
-      {:rule-name lint/let-if
-       :form (let [a 1] (if a (+ a a) 2))
+       :filename (io/file "corpus/throw_in_middle.clj")}
+      {:rule-name 'lint/let-if
+       :form '(let [a 1] (if a (+ a a) 2))
        :message "Use `if-let` instead of recreating it."
-       :alt (if-let [a 1] (+ a a) 2)
+       :alt '(if-let [a 1] (+ a a) 2)
        :line 9
        :column 1
        :end-col 29
        :end-row 9
-       :filename "corpus/throw_in_middle.clj"}]
+       :filename (io/file "corpus/throw_in_middle.clj")}]
     (io/file "corpus" "throw_in_middle.clj")))
 
 (defexpect parse-error-test
   (expect-match
-    '[{:rule-name splint/parsing-error
+    [{:rule-name 'splint/parsing-error
        :form nil
        :message "Splint encountered an error: Map literal contains duplicate key: :a"
        :alt nil
@@ -68,5 +68,5 @@
        :column 1
        :end-row nil
        :end-col nil
-       :filename "corpus/parse_error.clj"}]
+       :filename (io/file "corpus/parse_error.clj")}]
     (io/file "corpus" "parse_error.clj")))