+++ CHANGELOG.md
@@ -10,6 +10,13 @@ This changelog is loose. Versions are not semantic, they are incremental. Splint
 ### Added
 
 - `--[no]summary` cli flag to print or not print the summary line.
+- Extend the `matcher-combinators.core/Matcher` protocol to `java.io.File`, making `match?` work nicely with both strings and file objects.
+
+### Changed
+
+- `:filename` in `Diagnostic` is now a `java.io.File` object, not a string. This is propogated through everything. I suspect no one is using these so I think I could change the `Diagnostic` as well, but maybe I'll wait a min.
+- `make-edamame-opts` now accepts both `features` and `ns-state`, and `parse-string` and `parse-string-all` take in `features` instead of `ns-state`.
+- The runner produces objects instead of raw paths for processing. This includes which 
 
 ### Fixed